Extract override resolution logic from MissingOverrid rule
