+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/InternalApiBridge.java
@@ -143,6 +143,11 @@ public final class InternalApiBridge {
                 }
             });
     }
+    public static void overrideResolution(JavaAstProcessor processor, ASTCompilationUnit root) {
+        root.descendants(ASTAnyTypeDeclaration.class)
+            .crossFindBoundaries()
+            .forEach(OverrideResolutionPass::resolveOverrides);
+    }
 
     public static @Nullable JTypeMirror getTypeMirrorInternal(TypeNode node) {
         return ((AbstractJavaTypeNode) node).getTypeMirrorInternal();
